requirements.txt
worktrace
requests
python-dotenv
rich
